feat: support asynchronous api request handlers
